requirements.txt
chromadb>=0.4.24
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
redis>=5.0.4
redisvl>=0.2.2
pypdf>=4.2.0
numpy>=1.26.4
pandas>=2.1.4
networkx>=3.1
neo4j>=5.15.0
psutil>=5.9.8
tqdm>=4.66.5
python-dotenv>=1.0.0
tenacity>=8.3.0
scikit-learn>=1.3.2
sentence-transformers>=2.6.1
google-generativeai>=0.7.2

